Allow passing in of config filename to login server plugin
This means you can avoid editing a PHP file (and even store the list of
servers completely away from adminer.php if desired). Avoiding editing a
PHP file is desirable, as it could necessitate clearing an opcode cache
for the change to take effect.

This change was rejected upstream on the grounds that this could just as
easily be implemented by the code that instantiates the plugin. For now,
this change will remain in place, but I will look to handle this
differently in future.